refactor: moved all middlewares to `helpers/kindeMiddlewareHelpers.js`
