+++ src/index.js
@@ -1,48 +1,9 @@
-import { getInternalClient, setupInternalClient } from './setup';
+import { setupInternalClient } from './setup';
 import { setupAuthRouter } from './auth';
 
-const {
-  authToken,
-  getPem
-} = require("@kinde-oss/kinde-node-auth-utils").default;
+export * from './helpers';
 
-let pem;
-
-const setupKinde = async (config, app) => {
+export const setupKinde = async (config, app) => {
   setupInternalClient(app, config);
   setupAuthRouter(app, '/');
-  const { issuerBaseUrl } = config;
-  pem = await getPem(issuerBaseUrl);
-};
-
-const getUser = async (req, _, next) => { 
-  try {
-    const kindeClient = getInternalClient();
-    const userProfile = await kindeClient.getUserProfile(req);
-    req.user = userProfile;
-    next();
-  } catch (error) {
-    next(error);
-  }
-};
-
-const protectRoute = async (req, res, next) => {
-  const kindeClient = getInternalClient();
-  if (!await kindeClient.isAuthenticated(req)) {
-    return res.sendStatus(403);
-  }
-
-  const callback = (error) => {
-    if (error) return res.sendStatus(403);
-    next();
-  }
-
-  const parsedToken = await kindeClient.getToken(req);
-  authToken(parsedToken, pem, callback);
-};
-
-module.exports = {
-  setupKinde,
-  protectRoute,
-  getUser,
 };
